+++ CHANGELOG.md
@@ -15,6 +15,11 @@ For older change log history see the [historic changelog](HISTORIC_CHANGELOG.md)
     WebApplicationAuthenticationInformation in the schema and implementation.
     The schema referenced the class as DSC_WebApplicationAuthenticationInformation
     whereas the implementation referenced the class as MSFT_xWebApplicationAuthenticationInformation.
+- WebVirtualDirectory
+  [Issue #366](https://github.com/dsccommunity/WebAdministrationDsc/issues/366)
+  In WebVirtualDirectory WebApplication '' and '/' can now be used interchangeably.
+  - Fixed Add WebVirtualDirectory when WebApplication = '/'.
+  - Fixed Remove WebVirtualDirectory when WebApplication = ''.
 
 ## [4.0.0] - 2022-09-17